doc, test: fix type of NextSequence() keys
This functionality is implemented using `bolt.Bucket.NextSequence()`.
According to the [API docs][1], the value returned is a `uint64`.

The reason this was not detected by any tests and why it works with uint
is because of the use of `encoding/gob` which specifies that any
unsigned int type can be used to receive any unsigned int value provided
that it is large enough for the _actual_ value. Internally they all get
encoded at 64-bit values and the type is specific type is not recorded.

It also doesn't help that the tests have probably only ever been run on
64-bit systems where any uint64 value can be stored in a uint without
any errors if casting is used (like it is by `encoding/gob` internally).

[1]: https://godoc.org/github.com/boltdb/bolt#Bucket.NextSequence

+++ put_test.go
@@ -404,7 +404,7 @@ func TestInsertSequence(t *testing.T) {
 	testWrap(t, func(store *bolthold.Store, t *testing.T) {
 
 		type SequenceTest struct {
-			Key uint `boltholdKey:"Key"`
+			Key uint64 `boltholdKey:"Key"`
 		}
 
 		for i := 0; i < 10; i++ {
